requirements.txt
flask
torch
torchvision
ultralytics
pillow
opencv-python